+++ Logger/Entities/People/Employee.cs
@@ -2,9 +2,9 @@
 
 public record class Employee : Person
 {
-    public float Salary { get; init; }
+    public decimal Salary { get; init; }
 
-    public Employee(FullName fullName, float salary) : base(fullName)
+    public Employee(FullName fullName, decimal salary) : base(fullName)
     {
         if (salary <= 0) throw new ArgumentOutOfRangeException(nameof(salary), "Salary must be greater than 0");
         Salary = salary;
